requirements.txt
pyaudio
requests
openai-whisper
